# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codex3 API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/network-nodes/import:
    post:
      summary: Import socks5 or http proxy nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: One proxy per line
                default_protocol:
                  type: string
                  enum: [socks5, http]
                  description: Used when a line has no protocol prefix
                group_id:
                  type: string
                  format: uuid
                new_group_name:
                  type: string
      responses:
        "201":
          description: Import summary
...
  /api/v1/targets/import:
    post:
      summary: Import Telegram target links
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: One Telegram target link per line, for example https://t.me/AIGOGGGG
                group_id:
                  type: string
                  format: uuid
                new_group_name:
                  type: string
      responses:
        "201":
          description: Import summary
...